requirements.txt
# Python dependencies

# Core dependencies
llama-parse>=0.4.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.28.0

# Visualization and UI
streamlit>=1.28.0
plotly>=5.0.0
pandas>=2.0.0

# PDF processing
PyPDF2>=3.0.0
pdf2image>=1.16.0

# Development dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
